requirements.txt
python-telegram-bot[job-queue]==21.7
httpx==0.28.1
python-dotenv==1.0.1
